+++ src/data/constants.ts
@@ -3,7 +3,7 @@ import OriginIcons from "@/assets/OriginIcons";
 import WeaponIcons from "@/assets/WeaponIcons";
 import { computed } from "vue";
 import { DataSearchStrings } from "./services/dataSearchStringService";
-import type { IFilterButton, ILanguageInfo, IWeaponRangeValues } from "./interfaces";
+import { Collection, type ILanguageInfo, type IWeaponRangeValues, type SeasonNumber } from "./interfaces";
 
 export const LanguageInfos = computed<ILanguageInfo[]>(() => [
     { language: "de", flagIcon: FlagIcons.DE, text: "Altsächsisch", },
@@ -146,8 +146,9 @@ export const WeaponCategoryIconMap = computed<{ [itemRegex: string]: string }>((
     }
 });
 
+type SeasonIconMap = { [season in SeasonNumber]?: string };
 // TODO: custom icons for the older seasons
-export const SeasonIconMap = computed<{ [seasonNumber: number]: string }>(() => {
+export const SeasonIconMap = computed<SeasonIconMap>(() => {
     return {
         1: "",
         2: "",
@@ -160,30 +161,61 @@ export const SeasonIconMap = computed<{ [seasonNumber: number]: string }>(() => 
     }
 });
 
-// TODO: these filters
-export const OriginFilters = computed<IFilterButton[]>(() => [
-    { text: "World (Current)", iconUrl: "", filter: () => false, },
-    { text: "Word (Old)", iconUrl: "", filter: () => false, },
-    { text: "Vanguard Ops", iconUrl: OriginIcons.FactionVanguard, filter: () => false, },
-    { text: "Crucible", iconUrl: OriginIcons.FactionCrucible, filter: () => false, },
-    { text: "Gambit", iconUrl: OriginIcons.FactionGambit, filter: () => false, },
-    { text: "Iron Banner", iconUrl: OriginIcons.FactionIronBanner, filter: () => false, },
-    { text: "Trials of Osiris", iconUrl: OriginIcons.FactionOsiris, filter: () => false, },
-    { text: "Nightfall", iconUrl: OriginIcons.Nightfall, filter: () => false, },
-    { text: "King's Fall", iconUrl: "", filter: () => false, },
-    { text: "Duality", iconUrl: OriginIcons.Duality, filter: () => false, },
-    { text: "Opulent", iconUrl: "", filter: () => false, }, // TODO: this would be opulent weapons that were reissued ONLY (unless Include Sunset is checked?)
-    { text: "Vow of the Disciple", iconUrl: OriginIcons.VowOfTheDisciple, filter: () => false, },
-    { text: "Throne World", iconUrl: "", filter: () => false, },
-    { text: "30th Anniversary", iconUrl: "", filter: () => false, },
-    { text: "Vault of Glass", iconUrl: OriginIcons.VaultOfGlass, filter: () => false, },
-    { text: "Europa", iconUrl: OriginIcons.Europa, filter: () => false, },
-    { text: "Deep Stone Crypt", iconUrl: "", filter: () => false, },
-    { text: "Prophecy", iconUrl: OriginIcons.FactionTheNine, filter: () => false, }, // TODO: this would be Trials of the Nine weapons that drop in Prophecy ONLY (unless Include Sunset is checked?)
-    { text: "Altars of Sorrow", iconUrl: "", filter: () => false, },
-    { text: "Pit of Heresy", iconUrl: "", filter: () => false, },
-    { text: "Dreambane", iconUrl: "", filter: () => false, },
-    { text: "Garden of Salvation", iconUrl: "", filter: () => false, },
-    { text: "The Dreaming City", iconUrl: "", filter: () => false, },
-    { text: "Last Wish", iconUrl: "", filter: () => false, },
+interface IOriginFilterInfo {
+    collection: Collection;
+    text: string;
+    iconUrl: string;
+}
+
+type SeasonNumberToCollectionMap = { [season in SeasonNumber]: Collection };
+
+export const SeasonToCollectionMap = computed<SeasonNumberToCollectionMap>(() => {
+    return {
+        1: Collection.TheRedWar,
+        2: Collection.CurseOfOsiris,
+        3: Collection.Warmind,
+        4: Collection.Outlaw,
+        5: Collection.Forge,
+        6: Collection.Drifter,
+        7: Collection.Opulence,
+        8: Collection.Undying,
+        9: Collection.Dawn,
+        10: Collection.Worthy,
+        11: Collection.Arrivals,
+        12: Collection.Hunt,
+        13: Collection.Chosen,
+        14: Collection.Splicer,
+        15: Collection.Lost,
+        16: Collection.Risen,
+        17: Collection.Haunted,
+        18: Collection.Plunder,
+        19: Collection.Seraph,
+    }
+});
+
+export const OriginFilterInfos = computed<IOriginFilterInfo[]>(() => [
+    { collection: Collection.WorldCurrent, text: "World (Current)", iconUrl: "", },
+    { collection: Collection.WorldOld, text: "Word (Old)", iconUrl: "", },
+    { collection: Collection.VanguardOps, text: "Vanguard Ops", iconUrl: OriginIcons.FactionVanguard, },
+    { collection: Collection.Crucible, text: "Crucible", iconUrl: OriginIcons.FactionCrucible, },
+    { collection: Collection.Gambit, text: "Gambit", iconUrl: OriginIcons.FactionGambit, },
+    { collection: Collection.IronBanner, text: "Iron Banner", iconUrl: OriginIcons.FactionIronBanner, },
+    { collection: Collection.TrialsOfOsiris, text: "Trials of Osiris", iconUrl: OriginIcons.FactionOsiris, },
+    { collection: Collection.Nightfall, text: "Nightfall", iconUrl: OriginIcons.Nightfall, },
+    { collection: Collection.KingsFall, text: "King's Fall", iconUrl: "", },
+    { collection: Collection.Duality, text: "Duality", iconUrl: OriginIcons.Duality, },
+    { collection: Collection.Opulent, text: "Opulent", iconUrl: "", }, // TODO: this would be opulent weapons that were reissued ONLY (unless Include Sunset is checked?)
+    { collection: Collection.VowOfTheDisciple, text: "Vow of the Disciple", iconUrl: OriginIcons.VowOfTheDisciple, },
+    { collection: Collection.ThroneWorld, text: "Throne World", iconUrl: "", },
+    { collection: Collection.Anniversary, text: "30th Anniversary", iconUrl: "", },
+    { collection: Collection.VaultOfGlass, text: "Vault of Glass", iconUrl: OriginIcons.VaultOfGlass, },
+    { collection: Collection.Europa, text: "Europa", iconUrl: OriginIcons.Europa, },
+    { collection: Collection.DeepStoneCrypt, text: "Deep Stone Crypt", iconUrl: "", },
+    { collection: Collection.Prophecy, text: "Prophecy", iconUrl: OriginIcons.FactionTheNine, }, // TODO: this would be Trials of the Nine weapons that drop in Prophecy ONLY (unless Include Sunset is checked?)
+    { collection: Collection.AltarsOfSorrow, text: "Altars of Sorrow", iconUrl: "", },
+    { collection: Collection.PitOfHeresy, text: "Pit of Heresy", iconUrl: "", },
+    { collection: Collection.Dreambane, text: "Dreambane", iconUrl: "", },
+    { collection: Collection.GardenOfSalvation, text: "Garden of Salvation", iconUrl: "", },
+    { collection: Collection.TheDreamingCity, text: "The Dreaming City", iconUrl: "", },
+    { collection: Collection.LastWish, text: "Last Wish", iconUrl: "", },
 ]);