+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## **Name TBD** [ *1.1.1* ] Date TBD
+## **Name TBD** [ *1.2.0* ] Date TBD
 Description TBD
 
 * Fixed mods not having descriptions.
@@ -11,6 +11,7 @@ Description TBD
 * Sort weapon filters by name.
 * Sort weapon archetype filters by RPM/charge time if applicable.
 * Show Crafted bonus option now hidden for non-crafted weapons.
+* Added filters for seasonal weapons and various other collections (seasons prior to Season of Worthy not yet implemented).
 
 ## **Improve Page Load Times + Quality of Life** [ *1.1.0* ] 2023-02-13
 This change brings improvements to page load times as well as some quality of life changes such as